[refactor] feature 기반 clean architecture 구조에 맞게 폴더 구조 변경
