+++ test/widget/dialog_test.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:recipe_app/core/utils/app_show_dialog.dart';
-import 'package:recipe_app/presentation/widgets/base/button/app_button.dart';
+import 'package:recipe_app/core/presentation/widgets/button/app_button.dart';
 
 void main() {
   testWidgets('dialog test', (tester) async {
